Add linux fbdev display and keyboard backend
The mode is whatever was set when launching, though it would be trivial
to add support for mode setting.

Keyboard support is very primitive for now, basically it provides an
ESC key to kill the app and F1...F10 to console switch.

Note also that the fbdev is not switched to the front by default, so
you can setup your whole windowing system and only activate it all
at once when ready.