Add back support for pre-converted ttf fonts. The metrics generated by the converter might still be a bit dodgy, I'd appreciate somebody who knows that stuff better to have a look :-)
Basically, I added back the type '2' operation to the glyph data parser
that was used by the previous ttf parser before it got removed and fixed
it up to work in the new context. I pondered using the 'c' "generic"
curve operation instead and put the burden of conversion in the converter
application, but that would have caused bigger converted font files,
so I chose to keep the old op.

Various bits & pieces of the font code are modified to cope with the
non-stroke fonts (mostly there is no "pen" to account for and we don't
convolve but just append the glyph path to the current path).

Full support for unicode pages is back. The stroke font has only page 0
for now (the initial code seemed to have more tables I didn't feel like
re-converting that so if you want stroke font with more unicode pages,
it's up to you to dig the old format and convert it to detect snaps and
use curves). We may want to add arguments to twin_ttf converter to only
convert a subset of the pages in a font for compactness.

Examples still use the built-in font but it's easy enough to change the
g_twin_font global and link a generated font file in (tested with
Vera.ttf). I still need to add a better way to deal with multiple fonts
at once.