Fix timeout handling with more than one timeout
The code would do a delete/insert pass while walking the
queue which had the effect of breaking the "order" linkage
(insert re-initializes pretty much everything).

I fixed that by adding a new _twin_queue_reorder() which
is to be called on an element that is to be re-ordered and
which works at any time, even during a queue walking and
using it from the timeout code.


Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>