Add support for immediate damage refresh to fbdev
This makes the fbdev backend refresh damaged areas directly from
the screen->damaged callback instead of delayed to a "work". The
work is still there to handle console switches. In order to avoid
flicker and artifacts, we also disable refresh around cursor
updates and window content.

The demos flicker a lot in this mode, so they'll have to be fixed
to disable / enable refresh around some operations too (probably in
the widget code), but the overall performances are better for my
application as mouse cursor no longer triggers giant updates when
moving while something else is updating

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>