Add some basic altivec support
This adds altivec versions of argb32_over_argb32 and
argb32_source_argb32 operations and a bit of altivec
optimisation to the fbdev put span code as well.

It could probably be further optimized, and with some
patience, one could implement the full set of primitives
for altivec instead of just those two, but that's at
least a beginning. I haven't properly measured the
difference in performances yet. It is definitely faster
on my powerbook though.

Signed-off-by: Benjamin Herrenschmidt <benh@kernel.crashing.org>